tgBot2/requirements.txt
django==5.1.1 
psycopg2==2.9.9
python-dotenv==1.0.1
uvicorn==0.31.0
asyncio==3.4.3
pytelegrambotapi==4.23.0
aiohttp==3.10.9
logging==0.4.9.6
redis==5.1.1
setuptools==75.1.0
aio-pika==9.4.3
celery[aio]
geopy==2.4.1
pgeocode==0.5.0